(Hopefully) keep the block list correct after a write error.
If we're updating the original file, then we update the real_offset
after writing out a complete block, else we mark it as dirty to keep
it resident until we successfuly write it out.

If we're writing out a new file, leave the blocks alone until the
file has been completely written out, since the Buffer would still be
backed by the old file.