Imported prototype of relative value search.
This thing enables searching a file for a set of values, the
difference between each matches the difference between a provided
list of values.

A use for this is finding where offsets are stored in a file when
you do not know what the offsets may be relative to. A future
improvement could also be to handle factors of the search values
(e.g. offsets in units of >1 bytes).