Scale vertical scroll values when dealing with large files.
When dealing with very large files and/or very narrow windows, the
total number of lines in a Document can exceed the range of an int
and cause the scrolling code to get stuck in a loop or crash.

This limits the resolution of the vertical scrollbar to 10,000 steps
and divides the actual lines of the Document between them.