Check old_page_id isn't out of range in OnDocumentChange()
Due to a bug in wxWidgets, when adding a tab to a wxAuiNotebook under
very specific circumstances under OS X, extra page change events are
raised, with invalid old selection values. Additionally, the page
counter inside wxAuiNotebook is set to an out-of-range value, which
will still crash things if you call the wrong methods in the wrong
order...

I haven't managed to identify the root cause of the bug, and this
work around seems good enough to cover REHex's use of wxWidgets for
now.