Really re-open files in binary mode when saving on Windows.
Windows file descriptors can be opened in "text" or "binary"
mode, and the binary flag passed to fdopen() doesn't override
the file descriptor's implicit default of text mode.

When the file descriptor is opened in text mode, not only does
it do the usual newline mangling, it also considers a ^Z in the
file to be an EOF, which leads to us crashing from the fread()
failure in certain cases (this needs fixing too).