Optimise rendering of highlighted characters
This commit optimises the rendering of any characters with a coloured
background in the Document view.

Each area+line of non-hightlighted text was previously optimised into
a single wxDC::DrawText() call while any highlighted characters made
a call for each character. Each area+line+colour combination is now
handled using a single call, which improves performance when a lot of
text is selected or highlighted for any reason.