Initial work on 'DocumentCtrl' control.
This is mostly just a slightly gutted and mildly tweaked version of the
Document control at this point. The plan is for it to take over all
rendering functionality from Document, which can then be demoted from a
wxControl to a wxEvtHandler. Another class will act as glue code and
ferry updates back and forth between them. The control should become
flexible enough to use for the main view, side by side diff and
whatever comes next.

The DocumentCtrl will be responsible for handling only inputs that
affect its internal state (e.g. moving the cursor), anything further
such as editing the document will be propagated up to the glue code in
the parent class.