Mostly integrate DocumentCtrl properly.
This commit replaces the visible Document with DocumentCtrl, strips
most of the code from Document, removes DocumentCtrlTestWindow and
makes a lot of supporting changes to make everything mostly work.

Of note, copy and paste need reworking to fit the new model, some
display settings are still in Document pending removal and selection
needs to be fully removed from Document or tied in with DocumentCtrl
like the cursor position.