Use std::vector for ByteRangeSet storage.
Performance of std::vector is better than std::set overall for the
StringPanel uses.

Individual insertions are slower and bottleneck parallel searching for
strings, but nth element access is faster, as is adjusting for data
insert and erase operations, which caused the UI to hang when lots of
strings are present.