Implement ByteRangeSet::intersection() function.
This function makes a new ByteRangeSet containing the intersection of
two others. The current implementation won't scale to big data sets,
but the intended use is for moving ranges between dirty, pending and
processing queues in StringPanel, which should be dealing with tens of
ranges in each set at most.