Handle data being modified within a DisassemblyRegion.
Only handles overwrites since an insert or erase will result in the
regions being reinitialised anyway. Doesn't try to be clever with
re-inserting the new InstructionRegion and stitching it in with any
that were after it as processing is pretty fast anyway.

Updates of a processed region will not occur until the region is
visible on-screen... kind of sucks, but I don't want to add any
coupling that would allow the Region to call into DocumentCtrl to
tell it that it needs updating at this time.