Fix Windows build error?
I'm not even sure *WHY* this is an issue in the way it is.

wingdi.h defines an 'ERROR' macro (of course it does), which causes a
syntax error where ConsoleBuffer::Level::ERROR is used, except it
*doesn't* when I run the build myself - ONLY when running under
Buildkite.

I've spent well over an hour trying to figure out how/why Windows only
defines that macro under that EXACT situation - even copying the gcc
commands from the build log and running them on the exact same machine
by have doesn't reproduce it.

Sod it, this should make it build. Probably.