Refactor expression parsing
The nested branching expression grammars to handle operators
and their precedense doesn't scale. Adding any more operators
pushes it to unacceptable runtime even in the test samples.

So instead we parse expressions as a list of operands and
operators and compile them down to a tree ourselves. The
parser is only responsible for handling parens.

Seems to reduce the runtime by ~2/3rds on my laptop before
adding any more operators.