Assume any ASCII characters are fixed-width.
Another stupid problem caused by lying fonts - some of the
default fixed-width fonts on Windows claim certain plain ASCII
characters (like "m") are not the same width as others.

We used to assume (only) ASCII characters were safe to render
in the text view, so this assumption is probably okay.