Optimise file/line capture in template parser.
Doing a linear search through the file to find the file and line
number of each possible statement makes the parser obscenely slow.

Capture the input stream position instead and generate a sorted table
of line start/end positions to resolve it during post-processing.

This change drops the runtime from a little over 3 minutes for a ~7K
script down under a second.