Tweak array variable comments from binary templates.
When an array of POD values (i.e. int/etc) is declared, wrap the
whole array in a single comment so we don't waste the screen space
seperating every value and making strings harder to read.

An array of structs will still have a comment around each element
to group the members of each struct.