Initial implementation of template array resizing functions.
This is all going to have to be gutted and reworked... testing with
scripts from 010 Editor has proven the performance to be unusable
when there are arrays of thousands of POD values (e.g. char).

I've spent some time trying to optimise the array population, but I
suspect there's going to have to be some more abstraction built
into the data model, so that arrays of POD values can be flattened
down to a few objects... large arrays of structs may still prove to
be a problem.