Fix rendering of multibyte characters.
When drawing text, DocumentCtrl would search backwards from the
start of each line until it found a sequence that decoded as a
valid character, but in some encodings (e.g. Shift-JIS), this can
result in decoding garbage.

Instead... for encodings that don't mark the beginning/end of
multibyte sequences, we must decode all the way from the start
of the region. This is *SLOW*, so we do it in the background
and make an effort to cache the results, but it still has an
impact if you're wildly scrolling up and down a huge region
with those encodings.