Implement ByteRangeTree class to replace NestedOffsetLengthMap.
The NestedOffsetLengthMap container used for storing comments and
highlights doesn't scale up to large datasets - inserting requires a
linear search of all preceeding elements to check for invalid overlaps
since we only have a flat std::map to check.

Processing a NestedOffsetLength in a heirarchical manner is also
painful and so some code has to jump back and forth between iterators
to access elements in the desired order.

So, we now have a shiny new ByteRangeTree type to replace it!

This commit implements ByteRangeTree and guts most of the
NestedOffsetMap code, turning it into a wrapper around ByteRangeTree
so everything else doesn't need rewriting at once.

NestedOffsetLengthMap should be fully killed off and all dependant code
rewritten to use ByteRangeTree in the future.