Use faster iteration in REHex::Tab::compute_regions()
Now that the comments are in a proper tree strucutre, use it instead
of simulating a tree on top of the old NestedOffsetLengthKey iterators
which are simulated on top of the tree.

Improves performance somewhat and simplifies things.

Will probably factor out the actual tree walking code into either a
depth-first iterator helper class for ByteRangeTree in the future, or
change the behaviour of ByteRangeTree's own iterator to be depth-first
rather than insertion order.