[WIP] Reference variables via predefined stack addresses.
Allocate fixed (or function-relative) addresses for variables
instead of walking the stack to resolve them at each ref op.

Somehow this isn't actually a speed improvement, but is a necessary
prerequisite for resolving all types as a compilation step (due to the
way variables were previously scraped into struct scopes).