Do comments panel updates in batches (#205).
Populating the comments panel with a large number of comments can take
a while, most of which appears to be spent notifying the underlying
wxDataViewModel of changes, so we can't do it in a background thread.

We now do no more than 100 updates at a time, deferring further updates
into the idle event to keep the UI updating.