Add help page for value colourisation feature.
The Preferences dialog now has a context-sensitive "Help" button which
will display the manual page for the current page in the window.

Additionally, the Preferences dialog is now shown as a non-modal window
because trying to display a page via wxHtmlHelpController on wxGTK when
a modal dialog is open is broken (wxWidgets issue #24527).