Use GNU libiconv on macOS.
Apple managed to break their iconv in macOS 14, which was previously
causing the libunistring tests to fail, the solution is to build and
use libiconv from GNU rather than linking against the broken version
provided by Apple.

The CI workflow doesn't need to be pinned to macOS 13 any more with
this change.

https://savannah.gnu.org/bugs/index.php?65686
https://lists.gnu.org/archive/html/bug-gnulib/2024-05/msg00375.html