Fix race condition in ThreadPool.
My use of std::atomic<int> to implement a semaphore was fundamentally
flawed and has a race condition you can drive a bus through, yet it has
been a critical path in REHex for a while now and I can only trigger
it under a specific workload and only on my laptop for some reason.

Since every thread increments current_concurrency to check whether it
can proceed, a thread may increment the count, detect the overflow and
skip, but at the same time, depending on the whims of the scheduler,
a thread which was processing that task may decrement the counter and
then try to increment it again before the first thread decrements it,
leading to both/all threads determining there is no work to be done
and sleeping until new work is scheduled.