Save "bookmark" with security context for recent files on macOS.
When running under the application sandbox we can only access files chosen
by the user in the open/save dialogs, and restarting the application clears
the exceptions created that way.

Using "bookmarks", we can save the security context associated with each
file and restore the sandbox exceptions even after restarting.