Rework vcpkg installation process.
This commit switches vcpkg to "manifest mode" so that we can pin
dependencies to specific versions, as part of this, we have to also
pin the whole vcpkg tree at a specific version, so that gets updated
to the latest version when building a development snapshot.

The vcpkg install trees are now installed under the vcpkg-deploy
directory (per architecture) and will be cleaned like a normal VS
project.

See also:

https://github.com/microsoft/vcpkg/issues/46736
https://github.com/microsoft/vcpkg/pull/46473