Refactor tool panel drag implementation again (#252)
This commit decouples the overlay rendering details from the ToolDock
class and adds a THIRD overlay mechanism built on top of the wxOverlay
system, because the screen capture doesn't work under the KDE Wayland
display manager.

So, we use the screen capture overlay mechanism on Windows, where
blitting from the wxClientDC seems to work reliably, and improves UX on
earlier Windows versions where transparent windows don't work.

For Wayland, we use wxOverlay, which seems to work correctly with both
the GNOME and KDE display managers, but not anywhere else.

And finally, for X11 or macOS, we use the old popup window overlays.